+++ cmd/enabler/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/keitaroinc/enabler/cmd/enabler/app"
+)
+
+func main() {
+	app.Main()
+}
