+++ go.mod
@@ -17,7 +17,7 @@ require (
 	github.com/pkg/errors v0.8.1
 	github.com/sirupsen/logrus v1.4.1
 	github.com/spf13/cobra v1.0.0
-	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/spf13/pflag v1.0.5
 	github.com/spf13/viper v1.4.0
 	golang.org/x/sys v0.0.0-20200923182605-d9f96fdee20d // indirect
 	gopkg.in/src-d/go-git.v4 v4.13.1
